finished first prototype of custom packet paring
