+++ socketioxide/src/packet.rs
@@ -155,22 +155,10 @@ pub struct BinaryPacket {
     /// Binary payload
     pub bin: Vec<Vec<u8>>,
     /// The number of expected payloads (used when receiving data)
-    payload_count: usize,
+    /* todo: remove pub */ pub payload_count: usize,
 }
 
 impl<'a> PacketData<'a> {
-    fn index(&self) -> char {
-        match self {
-            PacketData::Connect(_) => '0',
-            PacketData::Disconnect => '1',
-            PacketData::Event(_, _, _) => '2',
-            PacketData::EventAck(_, _) => '3',
-            PacketData::ConnectError => '4',
-            PacketData::BinaryEvent(_, _, _) => '5',
-            PacketData::BinaryAck(_, _) => '6',
-        }
-    }
-
     /// Set the ack id for the packet
     /// It will only set the ack id for the packets that support it
     pub(crate) fn set_ack_id(&mut self, ack_id: i64) {
@@ -260,511 +248,8 @@ impl<'a> From<Packet<'a>> for String {
     }
 }
 
-/// Deserialize an event packet from a string, formated as:
-/// ```text
-/// ["<event name>", ...<JSON-stringified payload without binary>]
-/// ```
-fn deserialize_event_packet(data: &str) -> Result<(String, Value), Error> {
-    #[cfg(feature = "tracing")]
-    tracing::debug!("Deserializing event packet: {:?}", data);
-    let packet = match serde_json::from_str::<Value>(data)? {
-        Value::Array(packet) => packet,
-        _ => return Err(Error::InvalidEventName),
-    };
-
-    let event = packet
-        .first()
-        .ok_or(Error::InvalidEventName)?
-        .as_str()
-        .ok_or(Error::InvalidEventName)?
-        .to_string();
-    let payload = Value::from_iter(packet.into_iter().skip(1));
-    Ok((event, payload))
-}
-
-fn deserialize_packet<T: DeserializeOwned>(data: &str) -> Result<Option<T>, serde_json::Error> {
-    #[cfg(feature = "tracing")]
-    tracing::debug!("Deserializing packet: {:?}", data);
-    let packet = if data.is_empty() {
-        None
-    } else {
-        Some(serde_json::from_str(data)?)
-    };
-    Ok(packet)
-}
-
-/// Deserialize a packet from a string
-/// The string should be in the format of:
-/// ```text
-/// <packet type>[<# of binary attachments>-][<namespace>,][<acknowledgment id>][JSON-stringified payload without binary]
-/// + binary attachments extracted
-/// ```
-impl<'a> TryFrom<String> for Packet<'a> {
-    type Error = Error;
-
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        // It is possible to parse the packet from a byte slice because separators are only ASCII
-        let chars = value.as_bytes();
-        let mut i = 1;
-        let index = (b'0'..=b'6')
-            .contains(&chars[0])
-            .then_some(chars[0])
-            .ok_or(Error::InvalidPacketType)?;
-
-        // Move the cursor to skip the payload count if it is a binary packet
-        if index == b'5' || index == b'6' {
-            while chars.get(i) != Some(&b'-') {
-                i += 1;
-            }
-            i += 1;
-        }
-
-        let start_index = i;
-        // Custom nsps will start with a slash
-        let ns = if chars.get(i) == Some(&b'/') {
-            loop {
-                match chars.get(i) {
-                    Some(b',') => {
-                        i += 1;
-                        break Cow::Owned(value[start_index..i - 1].to_string());
-                    }
-                    // It maybe possible depending on clients that ns does not end with a comma
-                    // if it is the end of the packet
-                    // e.g `1/custom`
-                    None => {
-                        break Cow::Owned(value[start_index..i].to_string());
-                    }
-                    Some(_) => i += 1,
-                }
-            }
-        } else {
-            Cow::Borrowed("/")
-        };
-
-        let start_index = i;
-        let ack: Option<i64> = loop {
-            match chars.get(i) {
-                Some(c) if c.is_ascii_digit() => i += 1,
-                Some(b'[' | b'{') if i > start_index => break value[start_index..i].parse().ok(),
-                _ => break None,
-            }
-        };
-
-        let data = &value[i..];
-        let inner = match index {
-            b'0' => PacketData::Connect((!data.is_empty()).then(|| data.to_string())),
-            b'1' => PacketData::Disconnect,
-            b'2' => {
-                let (event, payload) = deserialize_event_packet(data)?;
-                PacketData::Event(event.into(), payload, ack)
-            }
-            b'3' => {
-                let packet = deserialize_packet(data)?.ok_or(Error::InvalidPacketType)?;
-                PacketData::EventAck(packet, ack.ok_or(Error::InvalidPacketType)?)
-            }
-            b'5' => {
-                let (event, payload) = deserialize_event_packet(data)?;
-                PacketData::BinaryEvent(event.into(), BinaryPacket::incoming(payload), ack)
-            }
-            b'6' => {
-                let packet = deserialize_packet(data)?.ok_or(Error::InvalidPacketType)?;
-                PacketData::BinaryAck(
-                    BinaryPacket::incoming(packet),
-                    ack.ok_or(Error::InvalidPacketType)?,
-                )
-            }
-            _ => return Err(Error::InvalidPacketType),
-        };
-
-        Ok(Self { inner, ns })
-    }
-}
-
 /// Connect packet sent by the client
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ConnectPacket {
-    sid: Sid,
-}
-
-#[cfg(test)]
-mod test {
-    use serde_json::json;
-
-    use super::*;
-
-    #[test]
-    fn packet_decode_connect() {
-        let sid = Sid::new();
-        let payload = format!("0{}", json!({ "sid": sid }));
-        let packet = Packet::try_from(payload).unwrap();
-
-        assert_eq!(Packet::connect("/", sid, ProtocolVersion::V5), packet);
-
-        let payload = format!("0/admin™,{}", json!({ "sid": sid }));
-        let packet = Packet::try_from(payload).unwrap();
-
-        assert_eq!(Packet::connect("/admin™", sid, ProtocolVersion::V5), packet);
-    }
-
-    #[test]
-    fn packet_encode_connect() {
-        let sid = Sid::new();
-        let payload = format!("0{}", json!({ "sid": sid }));
-        let packet: String = Packet::connect("/", sid, ProtocolVersion::V5)
-            .try_into()
-            .unwrap();
-        assert_eq!(packet, payload);
-
-        let payload = format!("0/admin™,{}", json!({ "sid": sid }));
-        let packet: String = Packet::connect("/admin™", sid, ProtocolVersion::V5)
-            .try_into()
-            .unwrap();
-        assert_eq!(packet, payload);
-    }
-
-    // Disconnect,
-
-    #[test]
-    fn packet_decode_disconnect() {
-        let payload = "1".to_string();
-        let packet = Packet::try_from(payload).unwrap();
-        assert_eq!(Packet::disconnect("/"), packet);
-
-        let payload = "1/admin™,".to_string();
-        let packet = Packet::try_from(payload).unwrap();
-        assert_eq!(Packet::disconnect("/admin™"), packet);
-    }
-
-    #[test]
-    fn packet_encode_disconnect() {
-        let payload = "1".to_string();
-        let packet: String = Packet::disconnect("/").try_into().unwrap();
-        assert_eq!(packet, payload);
-
-        let payload = "1/admin™,".to_string();
-        let packet: String = Packet::disconnect("/admin™").try_into().unwrap();
-        assert_eq!(packet, payload);
-    }
-
-    // Event(String, Value, Option<i64>),
-    #[test]
-    fn packet_decode_event() {
-        let payload = format!("2{}", json!(["event", { "data": "value" }]));
-        let packet = Packet::try_from(payload).unwrap();
-
-        assert_eq!(
-            Packet::event("/", "event", json!([{"data": "value"}])),
-            packet
-        );
-
-        // Check with ack ID
-        let payload = format!("21{}", json!(["event", { "data": "value" }]));
-        let packet = Packet::try_from(payload).unwrap();
-
-        let mut comparison_packet = Packet::event("/", "event", json!([{"data": "value"}]));
-        comparison_packet.inner.set_ack_id(1);
-        assert_eq!(packet, comparison_packet);
-
-        // Check with NS
-        let payload = format!("2/admin™,{}", json!(["event", { "data": "value™" }]));
-        let packet = Packet::try_from(payload).unwrap();
-
-        assert_eq!(
-            Packet::event("/admin™", "event", json!([{"data": "value™"}])),
-            packet
-        );
-
-        // Check with ack ID and NS
-        let payload = format!("2/admin™,1{}", json!(["event", { "data": "value™" }]));
-        let mut packet = Packet::try_from(payload).unwrap();
-        packet.inner.set_ack_id(1);
-
-        let mut comparison_packet = Packet::event("/admin™", "event", json!([{"data": "value™"}]));
-        comparison_packet.inner.set_ack_id(1);
-
-        assert_eq!(packet, comparison_packet);
-    }
-
-    #[test]
-    fn packet_encode_event() {
-        let payload = format!("2{}", json!(["event", { "data": "value™" }]));
-        let packet: String = Packet::event("/", "event", json!({ "data": "value™" }))
-            .try_into()
-            .unwrap();
-
-        assert_eq!(packet, payload);
-
-        // Encode empty data
-        let payload = format!("2{}", json!(["event", []]));
-        let packet: String = Packet::event("/", "event", json!([])).try_into().unwrap();
-
-        assert_eq!(packet, payload);
-
-        // Encode with ack ID
-        let payload = format!("21{}", json!(["event", { "data": "value™" }]));
-        let mut packet = Packet::event("/", "event", json!({ "data": "value™" }));
-        packet.inner.set_ack_id(1);
-        let packet: String = packet.try_into().unwrap();
-
-        assert_eq!(packet, payload);
-
-        // Encode with NS
-        let payload = format!("2/admin™,{}", json!(["event", { "data": "value™" }]));
-        let packet: String = Packet::event("/admin™", "event", json!({"data": "value™"}))
-            .try_into()
-            .unwrap();
-
-        assert_eq!(packet, payload);
-
-        // Encode with NS and ack ID
-        let payload = format!("2/admin™,1{}", json!(["event", { "data": "value™" }]));
-        let mut packet = Packet::event("/admin™", "event", json!([{"data": "value™"}]));
-        packet.inner.set_ack_id(1);
-        let packet: String = packet.try_into().unwrap();
-        assert_eq!(packet, payload);
-    }
-
-    // EventAck(Value, i64),
-    #[test]
-    fn packet_decode_event_ack() {
-        let payload = "354[\"data\"]".to_string();
-        let packet = Packet::try_from(payload).unwrap();
-
-        assert_eq!(Packet::ack("/", json!(["data"]), 54), packet);
-
-        let payload = "3/admin™,54[\"data\"]".to_string();
-        let packet = Packet::try_from(payload).unwrap();
-
-        assert_eq!(Packet::ack("/admin™", json!(["data"]), 54), packet);
-    }
-
-    #[test]
-    fn packet_encode_event_ack() {
-        let payload = "354[\"data\"]".to_string();
-        let packet: String = Packet::ack("/", json!("data"), 54).try_into().unwrap();
-        assert_eq!(packet, payload);
-
-        let payload = "3/admin™,54[\"data\"]".to_string();
-        let packet: String = Packet::ack("/admin™", json!("data"), 54)
-            .try_into()
-            .unwrap();
-        assert_eq!(packet, payload);
-    }
-
-    #[test]
-    fn packet_encode_connect_error() {
-        let payload = format!("4{}", json!({ "message": "Invalid namespace" }));
-        let packet: String = Packet::invalid_namespace("/").try_into().unwrap();
-        assert_eq!(packet, payload);
-
-        let payload = format!("4/admin™,{}", json!({ "message": "Invalid namespace" }));
-        let packet: String = Packet::invalid_namespace("/admin™").try_into().unwrap();
-        assert_eq!(packet, payload);
-    }
-
-    // BinaryEvent(String, BinaryPacket, Option<i64>),
-    #[test]
-    fn packet_encode_binary_event() {
-        let json = json!(["event", { "data": "value™" }, { "_placeholder": true, "num": 0}]);
-
-        let payload = format!("51-{}", json);
-        let packet: String =
-            Packet::bin_event("/", "event", json!({ "data": "value™" }), vec![vec![1]])
-                .try_into()
-                .unwrap();
-
-        assert_eq!(packet, payload);
-
-        // Encode with ack ID
-        let payload = format!("51-254{}", json);
-        let mut packet =
-            Packet::bin_event("/", "event", json!({ "data": "value™" }), vec![vec![1]]);
-        packet.inner.set_ack_id(254);
-        let packet: String = packet.try_into().unwrap();
-
-        assert_eq!(packet, payload);
-
-        // Encode with NS
-        let payload = format!("51-/admin™,{}", json);
-        let packet: String = Packet::bin_event(
-            "/admin™",
-            "event",
-            json!([{"data": "value™"}]),
-            vec![vec![1]],
-        )
-        .try_into()
-        .unwrap();
-
-        assert_eq!(packet, payload);
-
-        // Encode with NS and ack ID
-        let payload = format!("51-/admin™,254{}", json);
-        let mut packet = Packet::bin_event(
-            "/admin™",
-            "event",
-            json!([{"data": "value™"}]),
-            vec![vec![1]],
-        );
-        packet.inner.set_ack_id(254);
-        let packet: String = packet.try_into().unwrap();
-        assert_eq!(packet, payload);
-    }
-
-    #[test]
-    fn packet_decode_binary_event() {
-        let json = json!(["event", { "data": "value™" }, { "_placeholder": true, "num": 0}]);
-        let comparison_packet = |ack, ns: &'static str| Packet {
-            inner: PacketData::BinaryEvent(
-                "event".into(),
-                BinaryPacket {
-                    bin: vec![vec![1]],
-                    data: json!([{"data": "value™"}]),
-                    payload_count: 1,
-                },
-                ack,
-            ),
-            ns: ns.into(),
-        };
-
-        let payload = format!("51-{}", json);
-        let mut packet = Packet::try_from(payload).unwrap();
-        match packet.inner {
-            PacketData::BinaryEvent(_, ref mut bin, _) => bin.add_payload(vec![1]),
-            _ => (),
-        }
-
-        assert_eq!(packet, comparison_packet(None, "/"));
-
-        // Check with ack ID
-        let payload = format!("51-254{}", json);
-        let mut packet = Packet::try_from(payload).unwrap();
-        match packet.inner {
-            PacketData::BinaryEvent(_, ref mut bin, _) => bin.add_payload(vec![1]),
-            _ => (),
-        }
-
-        assert_eq!(packet, comparison_packet(Some(254), "/"));
-
-        // Check with NS
-        let payload = format!("51-/admin™,{}", json);
-        let mut packet = Packet::try_from(payload).unwrap();
-        match packet.inner {
-            PacketData::BinaryEvent(_, ref mut bin, _) => bin.add_payload(vec![1]),
-            _ => (),
-        }
-
-        assert_eq!(packet, comparison_packet(None, "/admin™"));
-
-        // Check with ack ID and NS
-        let payload = format!("51-/admin™,254{}", json);
-        let mut packet = Packet::try_from(payload).unwrap();
-        match packet.inner {
-            PacketData::BinaryEvent(_, ref mut bin, _) => bin.add_payload(vec![1]),
-            _ => (),
-        }
-        assert_eq!(packet, comparison_packet(Some(254), "/admin™"));
-    }
-
-    // BinaryAck(BinaryPacket, i64),
-    #[test]
-    fn packet_encode_binary_ack() {
-        let json = json!([{ "data": "value™" }, { "_placeholder": true, "num": 0}]);
-
-        let payload = format!("61-54{}", json);
-        let packet: String = Packet::bin_ack("/", json!({ "data": "value™" }), vec![vec![1]], 54)
-            .try_into()
-            .unwrap();
-
-        assert_eq!(packet, payload);
-
-        // Encode with NS
-        let payload = format!("61-/admin™,54{}", json);
-        let packet: String =
-            Packet::bin_ack("/admin™", json!({ "data": "value™" }), vec![vec![1]], 54)
-                .try_into()
-                .unwrap();
-
-        assert_eq!(packet, payload);
-    }
-
-    #[test]
-    fn packet_decode_binary_ack() {
-        let json = json!([{ "data": "value™" }, { "_placeholder": true, "num": 0}]);
-        let comparison_packet = |ack, ns: &'static str| Packet {
-            inner: PacketData::BinaryAck(
-                BinaryPacket {
-                    bin: vec![vec![1]],
-                    data: json!([{"data": "value™"}]),
-                    payload_count: 1,
-                },
-                ack,
-            ),
-            ns: ns.into(),
-        };
-
-        let payload = format!("61-54{}", json);
-        let mut packet = Packet::try_from(payload).unwrap();
-        match packet.inner {
-            PacketData::BinaryAck(ref mut bin, _) => bin.add_payload(vec![1]),
-            _ => (),
-        }
-
-        assert_eq!(packet, comparison_packet(54, "/"));
-
-        // Check with NS
-        let payload = format!("61-/admin™,54{}", json);
-        let mut packet = Packet::try_from(payload).unwrap();
-        match packet.inner {
-            PacketData::BinaryAck(ref mut bin, _) => bin.add_payload(vec![1]),
-            _ => (),
-        }
-
-        assert_eq!(packet, comparison_packet(54, "/admin™"));
-    }
-
-    #[test]
-    fn packet_size_hint() {
-        let sid = Sid::new();
-        let len = serde_json::to_string(&ConnectPacket { sid }).unwrap().len();
-        let packet = Packet::connect("/", sid, ProtocolVersion::V5);
-        assert_eq!(packet.get_size_hint(), len + 1);
-
-        let packet = Packet::connect("/admin", sid, ProtocolVersion::V5);
-        assert_eq!(packet.get_size_hint(), len + 8);
-
-        let packet = Packet::connect("admin", sid, ProtocolVersion::V4);
-        assert_eq!(packet.get_size_hint(), 8);
-
-        let packet = Packet::disconnect("/");
-        assert_eq!(packet.get_size_hint(), 1);
-
-        let packet = Packet::disconnect("/admin");
-        assert_eq!(packet.get_size_hint(), 8);
-
-        let packet = Packet::event("/", "event", json!({ "data": "value™" }));
-        assert_eq!(packet.get_size_hint(), 1);
-
-        let packet = Packet::event("/admin", "event", json!({ "data": "value™" }));
-        assert_eq!(packet.get_size_hint(), 8);
-
-        let packet = Packet::ack("/", json!("data"), 54);
-        assert_eq!(packet.get_size_hint(), 3);
-
-        let packet = Packet::ack("/admin", json!("data"), 54);
-        assert_eq!(packet.get_size_hint(), 10);
-
-        let packet = Packet::bin_event("/", "event", json!({ "data": "value™" }), vec![vec![1]]);
-        assert_eq!(packet.get_size_hint(), 3);
-
-        let packet = Packet::bin_event(
-            "/admin",
-            "event",
-            json!({ "data": "value™" }),
-            vec![vec![1]],
-        );
-        assert_eq!(packet.get_size_hint(), 10);
-
-        let packet = Packet::bin_ack("/", json!("data"), vec![vec![1]], 54);
-        assert_eq!(packet.get_size_hint(), 5);
-    }
+    pub sid: Sid,
 }
